requirements.txt
kafka-python==1.4.7
mysql-connector==2.2.9